Split the KotlinCompile.scala source file into multiple source files, each defining a single type definition
